Build 1.0.0.20 alpha
Fix: Filigree tooltips now clear previous tooltips effect descriptions before displaying.
Item Select Dialog: Weapons "Can Accept Sentience Jewel" flag now checked on an equipped item.
Breakdowns: Old weapon breakdowns removed.
Breakdowns: Weapon effect holder object linked up for weapon item selection.
Items: Test weapons/shields/orbs/runearms added. One of every type.
Items: First true weapon "Epic Bleeding Edge" added.
Engine: New notify UpdateGearChanged(slot) added.
Breakdowns: Dynamic breakdowns now added for items equipped in weapons slots. CURRENTLY ONLY SUPPORTS ACTUAL WEAPONS. SHIELDS, ORBS AND RUNEARMS TO COME LATER.
----Base damage bonus the X of X[1d8].
----Attack bonus.
-------Updated to list penalties due to TWF (if any).
----Damage bonus.
----Critical attack bonus.
-------Updated to list penalties due to TWF (if any).
----Critical damage bonus.
Items: Items can now have item specific effects that only apply to that item.
Fix: Changing the number of available build points now updates the ability +/- button states.
Fix: Feat tooltips for automatic and granted feats will no longer say the feat needs to be switched at Fred.
Items: Reinforced fists item effects added for all current items.
Breakdowns: Simple breakdown totals now shown to 2dp.
Enhancements: Legendary Dreadnought Masters Blitz effect descriptions shortened.
Enhancements: Kensei One with the Blade effect descriptions shortened.
Augments: "Legendary Slavelords Might: Melee Power" now has the correct text (Not ranged Power).
Stances: The following stances are now under program control. Based on equipped gear and relevant enhancements.
---Two weapon Fighting.
---Two handed fighting.
---Single weapon fighting.
------Bard "Swashbuckling Style - Skirmisher" enhancement allows use of a Buckler.
---Unarmed
---Sword and Board
---Staff
---Orb
---Rune Arm
---Swashbuckling
------Bard "Swashbuckling Style - Arcane Marauder" enhancement allows use of an Orb.
------Bard "Swashbuckling Style - Cannoneer" enhancement allows use of a Rune Arm.
------Bard "Swashbuckling Style - Skirmisher" enhancement allows use of a Buckler.
Equipment View: Equipping a two handed item in your main hand stops you from equipping anything in your off hand.
---Any off hand equipped item is removed.
Equipment View: Right clicking an item gives you the option of removing the equipped item.
UI: Descriptions of auto-controlled stances updated.
Stances View: User and Auto controlled stances are now separated in the UI.
UI: Level tooltip now includes the actual level it is for in its description text.
Fix: Fixed the resources to show correct version and product info on a properties click on the execuatble file.

+++ DDOCP/BreakdownItemSimple.cpp
@@ -29,8 +29,8 @@ CString BreakdownItemSimple::Value() const
 {
     CString value;
     value.Format(
-            "%3d",
-            (int)Total());
+            "%.2f",
+            Total());
     return value;
 }
